Como tarefa do issue OPENBUS-31: * Adicionando a linha de build necessária para compilar a OpenLDAP na Solaris 10, conforme bug indicado em: http://www.openldap.org/its/index.cgi/Build?id=4607. * Outras instruções de compilação na SunOS510
Como tarefa do issue OPENBUS-64:
* Criando um conjunto de testes para garantir que os scripts que rodam
os testes unitários bem como a recompilação em todas plataformas 
(via ssh), possam ser executados consistentemente. É interessante
reaproveitar essas verificações para os scripts de lançamento do 
openbus.
* Removendo a necessidade de definição da variável LUA51, agora o
packages.desc informa direto ao tecmake onde está o Lua
* BUGFIX: em algumas plataformas (solaris e alguns linux) o tar
reclama se não houver o parâmetro "-f -"

git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/trunk/tools@85947 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ recompile_all.sh
@@ -1,8 +1,47 @@
 #!/bin/ksh
 
+# helpers
+function is_dir {
+	dir=$1
+	if [ ! -d "$dir" ]; then
+		echo "ERROR: Missing directory $dir"
+		return 1
+	fi
+	return 0
+}
+function die {
+	echo $@
+	exit 1
+}
+
+TOOLSDIR=$OPENBUS_HOME/../trunk/tools
+LOGDIR=$OPENBUS_HOME/../
+
+# all checks before execute
+if [ "$OPENBUS_HOME" == "" ]
+then 
+	die "ERROR: Missing OPENBUS_HOME system variable"
+else
+	echo "INFO: Using OPENBUS_HOME as $OPENBUS_HOME"
+	# we don't test core/bin/$TEC_UNAME because this script could be run from any host
+	is_dir "$OPENBUS_HOME/core/bin" || die "ERROR: $OPENBUS_HOME seems an invalid OPENBUS_HOME"
+fi
+
+if [ ! -f $TOOLSDIR/compile.lua ]
+then
+	echo "INFO: We assume by default compile.lua on: $TOOLSDIR/compile.lua"
+	echo "INFO:                 and the log file on: $LOGDIR"
+	echo "INFO: But the compile.lua was not found!"
+	echo "INFO: Where is the compile.lua? Please inform us or CTRL+C to abort."
+	read NEWDIR
+	[  -f "$NEWDIR/compile.lua" ] || die "ERROR: Invalid directory $NEWDIR!"
+	echo "INFO: Thanks! Proceeding..."
+	TOOLSDIR=$NEWDIR
+fi
+
 timestamp=$(date +%Y%m%d)
-logfile="work/arch-built.$timestamp"
-compile_cmd="lua5.1 tools/compile.lua $@"
+logfile="$LOGDIR/arch-built.$timestamp"
+compile_cmd="cd $TOOLSDIR; ./compile.lua $@"
 
 rm -f $logfile
 touch $logfile
